refactor: download binary on demand
