+++ package.json
@@ -7,7 +7,6 @@
   },
   "files": [
     "index.js",
-    "install.js",
     "bin/",
     "LICENSE",
     "README.md"
@@ -17,9 +16,6 @@
     "url": "https://github.com/mochilang/mochi"
   },
   "license": "MIT",
-  "scripts": {
-    "postinstall": "node install.js"
-  },
   "dependencies": {
     "tar": "^6.2.0",
     "tree-sitter": "^0.22.4",
